+++ osidb/models/tests/test_affect.py
@@ -1,11 +1,13 @@
 import pytest
+from django.core.exceptions import ValidationError
 
-from osidb.models import Affect, Impact
+from osidb.models import Affect, Impact, Tracker
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
     PsModuleFactory,
     PsUpdateStreamFactory,
+    TrackerFactory,
 )
 
 pytestmark = pytest.mark.unit
@@ -149,3 +151,53 @@ class TestAffect:
 
         # Check that the label was automatically added to the labels field
         assert "test-auto-label" in affect.labels
+
+    def test_delete_affect_with_open_tracker(self):
+        """Test that an affect with a non closed tracker can't be deleted"""
+        ps_module = PsModuleFactory(bts_name="jboss")
+        ps_update_stream = PsUpdateStreamFactory(ps_module=ps_module)
+        flaw = FlawFactory(embargoed=False)
+
+        affect = AffectFactory(
+            flaw=flaw,
+            affectedness=Affect.AffectAffectedness.AFFECTED,
+            resolution=Affect.AffectResolution.DELEGATED,
+            ps_update_stream=ps_update_stream.name,
+            ps_module=ps_module.name,
+            ps_component="test-component",
+        )
+
+        tracker = TrackerFactory(
+            embargoed=False,
+            affects=[affect],
+            type=Tracker.TrackerType.JIRA,
+            external_system_id="OSIDB-1234",
+            ps_update_stream=ps_update_stream.name,
+            status="Open",
+        )
+
+        # Trying to delete the affect without first closing the Tracker should raise an error
+        with pytest.raises(ValidationError):
+            affect.delete()
+
+        tracker.status = "Closed"
+        tracker.save()
+
+        # Tracker is closed, should not raise an error
+        affect.delete()
+
+    def test_delete_affect_without_tracker(self):
+        """Test that an affect without a tracker can be deleted"""
+        ps_module = PsModuleFactory(bts_name="jboss")
+        ps_update_stream = PsUpdateStreamFactory(ps_module=ps_module)
+        flaw = FlawFactory(embargoed=False)
+
+        affect = AffectFactory(
+            flaw=flaw,
+            affectedness=Affect.AffectAffectedness.AFFECTED,
+            resolution=Affect.AffectResolution.DELEGATED,
+            ps_update_stream=ps_update_stream.name,
+            ps_module=ps_module.name,
+            ps_component="test-component",
+        )
+        affect.delete()